+++ render.c
@@ -231,6 +231,13 @@ render_prompt(const struct render *render, struct buffer *buf,
     char32_t prev = 0;
 
     for (size_t i = 0; i < prompt_len + text_len; i++) {
+        if (i >= prompt_len &&
+            conf->password_mode.enabled &&
+            conf->password_mode.character == U'\0')
+        {
+            continue;
+        }
+
         char32_t wc = i < prompt_len
             ? pprompt[i]
             : (conf->password_mode.enabled